+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- `Select-DataJourney` failed when selecting models by key instead of title.
+
 ## [0.4.0] - 2024-11-19
 
 ### Added